Refactored local notifications in separate category
